preprocess/requirements.txt
bs4
zenhan
mecab-python3
pyknp
